create sso-saml module. Move all saml related files (services, controllers, tests, helpers) into the module
