+++ packages/@n8n/backend-common/src/modules/modules.config.ts
@@ -14,6 +14,7 @@ export const MODULE_NAMES = [
 	'dynamic-credentials',
 	'chat-hub',
 	'sso-oidc',
+	'sso-saml',
 	'log-streaming',
 ] as const;
