+++ packages/cli/src/modules/sso-saml/schema/metadata-exchange.xsd.ts
