+++ packages/cli/src/modules/sso-saml/schema/oasis-200401-wss-wssecurity-utility-1.0.xsd.ts
