+++ packages/cli/src/modules/sso-saml/schema/saml-schema-assertion-2.0.xsd.ts
