+++ packages/cli/src/modules/sso-saml/schema/saml-schema-metadata-2.0.xsd.ts
