+++ packages/cli/src/modules/sso-saml/schema/saml-schema-protocol-2.0.xsd.ts
