+++ packages/cli/src/modules/sso-saml/schema/ws-addr.xsd.ts
