+++ packages/cli/src/modules/sso-saml/schema/ws-authorization.xsd.ts
