+++ packages/cli/src/modules/sso-saml/schema/ws-federation.xsd.ts
