+++ packages/cli/src/modules/sso-saml/schema/ws-securitypolicy-1.2.xsd.ts
