+++ packages/cli/src/modules/sso-saml/schema/xml.xsd.ts
