+++ packages/cli/src/modules/sso-saml/schema/xmldsig-core-schema.xsd.ts
