+++ packages/cli/src/sso.ee/sso-helpers.ts
@@ -1,9 +1,9 @@
+import { LicenseState, Logger } from '@n8n/backend-common';
 import { GlobalConfig } from '@n8n/config';
 import { isAuthProviderType, SettingsRepository, type AuthProviderType } from '@n8n/db';
 import { Container } from '@n8n/di';
 
 import config from '@/config';
-import { Logger } from '@n8n/backend-common';
 
 /**
  * Only one authentication method can be active at a time. This function sets
@@ -50,6 +50,25 @@ export function getCurrentAuthenticationMethod(): AuthProviderType {
 
 export function isSamlCurrentAuthenticationMethod(): boolean {
 	return getCurrentAuthenticationMethod() === 'saml';
+}
+
+/**
+ *  Check whether the SAML feature is licensed and enabled in the instance
+ */
+export function isSamlLoginEnabled(): boolean {
+	return Container.get(GlobalConfig).sso.saml.loginEnabled;
+}
+
+export function getSamlLoginLabel(): string {
+	return Container.get(GlobalConfig).sso.saml.loginLabel;
+}
+
+export function isSamlLicensed(): boolean {
+	return Container.get(LicenseState).isSamlLicensed();
+}
+
+export function isSamlLicensedAndEnabled(): boolean {
+	return isSamlLoginEnabled() && isSamlLicensed() && isSamlCurrentAuthenticationMethod();
 }
 
 export function isLdapCurrentAuthenticationMethod(): boolean {
