+++ packages/cli/test/integration/saml/saml-helpers.test.ts
@@ -1,7 +1,7 @@
 import { getPersonalProject, testDb } from '@n8n/backend-test-utils';
 
-import * as helpers from '@/sso.ee/saml/saml-helpers';
-import type { SamlUserAttributes } from '@/sso.ee/saml/types';
+import * as helpers from '@/modules/sso-saml/saml-helpers';
+import type { SamlUserAttributes } from '@/modules/sso-saml/types';
 
 beforeAll(async () => {
 	await testDb.init();
