+++ packages/cli/test/integration/shared/utils/test-server.ts
@@ -217,10 +217,10 @@ export const setupTestServer = ({
 					}
 
 					case 'saml': {
-						const { SamlService } = await import('@/sso.ee/saml/saml.service.ee');
+						const { SamlService } = await import('@/modules/sso-saml/saml.service.ee');
 						await Container.get(SamlService).init();
-						await import('@/sso.ee/saml/routes/saml.controller.ee');
-						const { setSamlLoginEnabled } = await import('@/sso.ee/saml/saml-helpers');
+						await import('@/modules/sso-saml/saml.controller.ee');
+						const { setSamlLoginEnabled } = await import('@/modules/sso-saml/saml-helpers');
 						await setSamlLoginEnabled(true);
 						break;
 					}
